Renamed directories to have same name as module. Applied to: maven-kbcsv-plugin, maven-testgenerator-plugin, rules-testing-tools (all 3 now prefixed with "asqatasun-")
