+++ testing-tools/asqatasun-maven-testgenerator-plugin/src/main/java/org/asqatasun/test/generator/GeneratorException.java
