+++ testing-tools/asqatasun-rules-testing-tools/src/main/resources/context/beans-contentadapter.xml
