+++ testing-tools/asqatasun-rules-testing-tools/src/main/resources/context/beans-contentloader.xml
