+++ testing-tools/asqatasun-rules-testing-tools/src/main/resources/context/beans-persistence.xml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:p="http://www.springframework.org/schema/p" xmlns:aop="http://www.springframework.org/schema/aop"
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:tx="http://www.springframework.org/schema/tx"
+       xsi:schemaLocation="http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.0.xsd
+       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+       http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd
+
+       http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-3.0.xsd">
+
+    <bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
+        <property name="dataSource" ref="dataSource"/>
+        <property name="persistenceXmlLocation" value="classpath:context/persistence.xml"/>
+        <property name="jpaProperties">
+            <props>
+                <prop key="hibernate.show_sql">false</prop>
+            </props>
+        </property>
+    </bean>
+    <bean id="dataSource"
+          class="org.springframework.jdbc.datasource.DriverManagerDataSource"
+          p:driverClassName="org.hsqldb.jdbcDriver"
+          p:url="jdbc:hsqldb:file:src/main/resources/hsql-db"
+          p:username="sa"
+          p:password=""
+          />
+    <!--
+        PostProcessors to perform exception translation on @Repository classes (from native
+        exceptions such as JPA PersistenceExceptions to Spring's DataAccessException hierarchy).
+        -->
+    <bean class="org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor"/>
+    <bean class="org.springframework.orm.jpa.support.PersistenceAnnotationBeanPostProcessor"/>
+    <!-- Transaction manager for  JTA  -->
+    <!-- bean id="transactionManager" class="org.springframework.transaction.jta.JtaTransactionManager" -->
+    <bean id="transactionManager" class="org.springframework.orm.jpa.JpaTransactionManager">
+        <property name="entityManagerFactory" ref="entityManagerFactory"/>
+    </bean>
+    <!--
+        Instruct Spring to perform declarative transaction management
+        automatically on annotated classes.
+        -->
+    <!-- tx:annotation-driven / -->
+    <!--
+        Activates various annotations to be detected in bean classes: Spring's
+        @Required and @Autowired, as well as JSR 250's @PostConstruct,
+        @PreDestroy and @Resource (if available) and JPA's @PersistenceContext
+        and @PersistenceUnit (if available).
+        -->
+    <context:annotation-config/>
+
+    <aop:config>
+        <aop:pointcut id="transactionalMethods" expression="bean(*DAO) || bean(*Service) || bean(eaLoggerComponent)"/>
+        <aop:advisor advice-ref="txAdvice" pointcut-ref="transactionalMethods"/>
+    </aop:config>
+
+    <tx:advice id="txAdvice" transaction-manager="transactionManager">
+        <tx:attributes>
+            <tx:method name="create*" propagation="REQUIRED"/>
+            <tx:method name="update*" propagation="REQUIRED"/>
+            <tx:method name="delete*" propagation="REQUIRED"/>
+            <!--<tx:method name="read*" propagation="REQUIRED"/>-->
+            <tx:method name="save*" propagation="REQUIRED"/>
+
+            <tx:method name="init" propagation="REQUIRED"/>
+            <tx:method name="crawl*" propagation="REQUIRED"/>
+            <tx:method name="load*" propagation="REQUIRED"/>
+            <tx:method name="adapt*" propagation="REQUIRED"/>
+            <tx:method name="process*" propagation="REQUIRED"/>
+            <tx:method name="consolidate*" propagation="REQUIRED"/>
+            <tx:method name="analyse*" propagation="REQUIRED"/>
+
+            <!--Sans ça, tout pète-->
+            <!--<tx:method name="audit*" propagation="REQUIRED"/>-->
+            <!--<tx:method name="*" propagation="SUPPORTS" read-only="true"/>-->
+        </tx:attributes>
+    </tx:advice>
+
+</beans>
+
